refactor: SFSymbol, +SwiftUI, SymbolButtons
