+++ SwiftDiffusion/Views/Symbols/SFSymbol.swift
@@ -1,0 +1,17 @@
+//
+//  SFSymbol.swift
+//  SwiftDiffusion
+//
+//  Created by Justin Bush on 3/8/24.
+//
+
+import Foundation
+
+enum SFSymbol: String {
+  case coreMl = "rotate.3d"
+  case python = "point.bottomleft.forward.to.point.topright.scurvepath"
+  case trash
+  case lock
+  
+  case none
+}
