+++ SwiftDiffusion/Views/WindowViews/CheckpointManagerView/CheckpointManagerView.swift
@@ -158,51 +158,3 @@ extension CheckpointsManager {
     }
   }
 }
-
-enum SFSymbol: String {
-  case coreMl = "rotate.3d"
-  case python = "point.bottomleft.forward.to.point.topright.scurvepath"
-  case trash
-  case lock
-  
-  case none
-  
-  var name: String {
-    return self.rawValue
-  }
-  
-  var image: some View {
-    Image(systemName: self.name)
-  }
-}
-
-struct MenuButton: View {
-  let title: String
-  var symbol: SFSymbol = .none
-  let action: () -> Void
-  
-  var body: some View {
-    Button(action: action) {
-      HStack {
-        if symbol != .none {
-          symbol.image
-        }
-        Text(title)
-      }
-    }
-  }
-}
-
-struct SymbolButton: View {
-  let symbol: SFSymbol
-  let action: () -> Void
-  
-  var body: some View {
-    Button(action: {
-      action()
-    }) {
-      symbol.image
-    }
-    .buttonStyle(BorderlessButtonStyle())
-  }
-}
